Add ExportAnnotationsActor
