+++ akka-bbb-apps/src/main/scala/org/bigbluebutton/core/record/events/StoreAnnotationsInRedisPresAnnEvent.scala
@@ -22,51 +22,27 @@ package org.bigbluebutton.core.record.events
 import org.bigbluebutton.common2.msgs.{ AnnotationVO, ExportJob, StoredAnnotations, PresentationPageForExport }
 import org.bigbluebutton.common2.util.JsonUtil
 
-class StoreExportJobInRedisRecordEvent extends AbstractPresentationWithAnnotations {
-  import StoreExportJobInRedisRecordEvent._
+class StoreAnnotationsInRedisPresAnnEvent extends AbstractPresentationWithAnnotations {
+  import StoreAnnotationsInRedisPresAnnEvent._
 
-  setEvent("StoreExportJobInRedisRecordEvent")
+  setEvent("StoreAnnotationsInRedisPresAnnEvent")
 
   def setJobId(jobId: String) {
     eventMap.put(JOB_ID, jobId)
-  }
-
-  def setJobType(jobType: String) {
-    eventMap.put(JOB_TYPE, jobType)
   }
 
   def setPresId(presId: String) {
     eventMap.put(PRES_ID, presId)
   }
 
-  def setPresLocation(presLocation: String) {
-    eventMap.put(PRES_LOCATION, presLocation)
-  }
-
-  def setAllPages(allPages: String) {
-    eventMap.put(ALL_PAGES, allPages)
-  }
-
   def setPages(pages: Array[PresentationPageForExport]) {
     eventMap.put(PAGES, JsonUtil.toJson(pages))
   }
-
-  def setParentMeetingId(parentMeetingId: String) {
-    eventMap.put(PARENT_MEETING_ID, parentMeetingId)
-  }
-
-  def setPresentationUploadToken(presentationUploadToken: String) {
-    eventMap.put(PRESENTATION_UPLOAD_TOKEN, presentationUploadToken)
-  }
 }
 
-object StoreExportJobInRedisRecordEvent {
+object StoreAnnotationsInRedisPresAnnEvent {
   protected final val JOB_ID = "jobId"
-  protected final val JOB_TYPE = "jobType"
   protected final val PRES_ID = "presId"
-  protected final val PRES_LOCATION = "presLocation"
-  protected final val ALL_PAGES = "allPages"
   protected final val PAGES = "pages"
-  protected final val PARENT_MEETING_ID = "parentMeetingId"
-  protected final val PRESENTATION_UPLOAD_TOKEN = "presentationUploadToken"
 }
+
