+++ akka-bbb-apps/src/main/scala/org/bigbluebutton/endpoint/redis/RedisRecorderActor.scala
@@ -49,10 +49,6 @@ class RedisRecorderActor(
     redis.recordAndExpire(session, message)
   }
 
-  private def storePresentationAnnotations(session: String, message: java.util.Map[java.lang.String, java.lang.String], messageType: String): Unit = {
-    redis.storePresentationAnnotations(session, message, messageType)
-  }
-
   def receive = {
     //=============================
     // 2x messages
@@ -77,8 +73,6 @@ class RedisRecorderActor(
       case m: CreateNewPresentationPodEvtMsg        => handleCreateNewPresentationPodEvtMsg(m)
       case m: RemovePresentationPodEvtMsg           => handleRemovePresentationPodEvtMsg(m)
       case m: SetPresenterInPodRespMsg              => handleSetPresenterInPodRespMsg(m)
-      case m: StoreAnnotationsInRedisSysMsg         => handleStoreAnnotationsInRedisSysMsg(m)
-      case m: StoreExportJobInRedisSysMsg           => handleStoreExportJobInRedisSysMsg(m)
 
       // Whiteboard
       case m: SendWhiteboardAnnotationEvtMsg        => handleSendWhiteboardAnnotationEvtMsg(m)
@@ -137,31 +131,6 @@ class RedisRecorderActor(
 
       case _                                        => // message not to be recorded.
     }
-  }
-
-  private def handleStoreAnnotationsInRedisSysMsg(msg: StoreAnnotationsInRedisSysMsg) {
-    val ev = new StorePresentationAnnotationsRecordEvent()
-
-    ev.setJobId(msg.body.annotations.jobId)
-    ev.setPresId(msg.body.annotations.presId)
-    ev.setPages(msg.body.annotations.pages)
-
-    storePresentationAnnotations(msg.header.meetingId, ev.toMap.asJava, "Annotations")
-  }
-
-  private def handleStoreExportJobInRedisSysMsg(msg: StoreExportJobInRedisSysMsg) {
-    val ev = new StoreExportJobInRedisRecordEvent()
-
-    ev.setJobId(msg.body.exportJob.jobId)
-    ev.setJobType(msg.body.exportJob.jobType)
-    ev.setPresId(msg.body.exportJob.presId)
-    ev.setPresLocation(msg.body.exportJob.presLocation)
-    ev.setAllPages(msg.body.exportJob.allPages.toString)
-    ev.setPages(msg.body.exportJob.pages)
-    ev.setParentMeetingId(msg.body.exportJob.parentMeetingId)
-    ev.setPresentationUploadToken(msg.body.exportJob.presUploadToken)
-
-    storePresentationAnnotations(msg.header.meetingId, ev.toMap.asJava, "ExportJob")
   }
 
   private def handleGroupChatMessageBroadcastEvtMsg(msg: GroupChatMessageBroadcastEvtMsg) {
